Add administrative events overview
Add the possibility for admins to view
a) All events for a given time span in a feed-like manner
b) An histogram of all events of the last n days (def. 1 year)

Currently the general idea is a GitHub-like overview
of everything that happened in the last year.

This is the server-side work corresponding to
tradity/tradity-client#21.